Feat: Relevant jsx turned into 'FileUploader' component.
